pyproject.toml
[tool.poetry]
name = "meta-ads-bigquery"
version = "3.6"
description = "A script to load Meta Ads data into BigQuery"
readme = "README.md"
authors = ["Yehor Smoliakov <yehors+gh@ukr.net>"]
classifiers = [
    "Programming Language :: Python",
    'Programming Language :: Python :: 3',
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
facebook-business = "==18.0.4"
"google-cloud-bigquery" = "==3.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
